Splete row_composite_serializer.py into several smaller
