+++ src/django/contricleaner/lib/row_serializers/row_cleaned_user_data_serializer.py
@@ -1,0 +1,4 @@
+class RowCleanedUserDataSerializer:
+    def validate(self, row: dict, current: dict) -> dict:
+        current["cleaned_user_data"] = row.copy()
+        return current
