+++ src/django/contricleaner/lib/row_serializers/row_composite_serializer.py
@@ -1,61 +1,17 @@
 from contricleaner.lib.dto.row_dto import RowDTO
-from contricleaner.lib.helpers.clean import clean
-from countries.lib.get_country_code import get_country_code
 
-
-class RowSectorSerializer:
-    def validate(self, row: dict, current: dict) -> dict:
-        current["sector"] = row["sector"].split(",")
-
-
-class RowCleanFieldSerializer:
-    def __init__(self, field: str, new_field: str) -> None:
-        self.field = field
-        self.new_field = new_field
-
-    def validate(self, row: dict, current: dict) -> dict:
-        clean_value = clean(row[self.field])
-        if len(clean_value) == 0:
-            current["errors"].append(
-                {
-                    "message": "{} cannot be empty".format(self.new_field),
-                    "type": "Error",
-                }
-            )
-            return current
-
-        current[self.new_field] = clean_value
-        return current
-
-
-class RowCleanedUserDataSerializer:
-    def validate(self, row: dict, current: dict) -> dict:
-        current["cleaned_user_data"] = row.copy()
-        return current
-
-
-class RowEmptySerializer:
-    def validate(self, row: dict, current: dict) -> dict:
-        keys = set(row.keys()).difference(current.keys())
-        for key in keys:
-            current[key] = row[key]
-
-        return current
-
-
-class RowCountrySerializer:
-    def validate(self, row: dict, current: dict) -> dict:
-        try:
-            current["country_code"] = get_country_code(row["country"])
-            return current
-        except ValueError as exc:
-            current["errors"].append(
-                {
-                    "message": str(exc),
-                    "type": "Error",
-                }
-            )
-            return current
+from django.contricleaner.lib.row_serializers.row_clean_field_serializer import (
+    RowCleanFieldSerializer,
+)
+from django.contricleaner.lib.row_serializers.row_country_serializer import (
+    RowCountrySerializer,
+)
+from django.contricleaner.lib.row_serializers.row_empty_serializer import (
+    RowEmptySerializer,
+)
+from django.contricleaner.lib.row_serializers.row_sector_serializer import (
+    RowSectorSerializer,
+)
 
 
 class RowCompositeSerializer:
@@ -90,6 +46,7 @@ class RowCompositeSerializer:
         dict_res = {
             "fields": {},
         }
+
         for res_key, res_value in res.items():
             if res_key in standard_fields:
                 dict_res[res_key] = res_value
