+++ src/django/contricleaner/lib/row_serializers/row_empty_serializer.py
@@ -1,0 +1,7 @@
+class RowEmptySerializer:
+    def validate(self, row: dict, current: dict) -> dict:
+        keys = set(row.keys()).difference(current.keys())
+        for key in keys:
+            current[key] = row[key]
+
+        return current
