+++ src/django/contricleaner/lib/row_serializers/row_sector_serializer.py
@@ -1,0 +1,3 @@
+class RowSectorSerializer:
+    def validate(self, row: dict, current: dict) -> dict:
+        current["sector"] = row["sector"].split(",")
